+++ docs/source/changes.md
@@ -45,6 +45,11 @@ authors:
   processing. This feature is enabled by default and can be switched off via
   the [`find_breaks`][config.find_breaks] setting.
   ({{ gh(386) }} by {{ authors.hoechenberger }})
+- You can now use the FreeSurfer `fsaverage` template MRI for source estimation
+  in cases where you don't have participant-specific MR scans available, as is
+  often the case in EEG studies. The behavior can be enabled using the new
+  configuration option [`use_template_mri`][config.use_template_mri].
+  ({{ gh(387) }} by {{ authors.agramfort }} and {{authors.hoechenberger }})
 
 ### Behavior changes